refactor: Move domain to Form\LegacyConsumer.
