+++ src/Form/LegacyConsumer/Commands/HookCommandInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\FieldsAPI\Commands;
+namespace Give\Form\LegacyConsumer\Commands;
 
 interface HookCommandInterface {
     public function __invoke( $hook );
