+++ src/Form/LegacyConsumer/FilterCallbackCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\FieldsAPI;
+namespace Give\Form\LegacyConsumer;
 
 class FilterCallbackCollection {
